src/requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
fastapi==0.81.0
uvicorn
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python-headless>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
torchvision==0.12.0+cpu
tqdm>=4.41.0

# plotting ------------------------------------

pandas>=1.1.4
seaborn>=0.11.0

# deep_sort -----------------------------------

easydict

# torchreid

Cython
h5py
six
tb-nightly
future
yacs
gdown
flake8
yapf
isort==4.3.21
imageio

mysql-connector-python
youtube_dl
datetime
ImageHash
python-dotenv